Polish buildpack configuration
This change polishes up the buildpack configuration.  Rather than
having multiple environment variables that can affect which buildpack
is used, only a single variable is now exposed; CF_BUILDPACK.  By
default, this is set to https://github.com/cloudfoundry/java-
buildpack.git, but it can be overridden when needed.  This change
causes the system-test harness to ignore the buildpack specified in
the java-test-applications manifest.yml file.