Make database tests insensitive to the query part of the uri
There is a diffrence between spring-cloud and cloudfoundry-runtime on
how database url is created (and passed on the the driver). Transitively,
this affects auto-reconfig based on those libraries. This change removes
the query parameters from the url obtained throught REST call. This makes
the tests pass auto-reconfig tests before and after using spring-cloud.