Add tests and extract out JDBC URL normalization
This change adds tests for the new (and changed) code that was added
to support testing of Postgres auto-reconfiguration.  The algorithm
used to normalize the JDBC URL, both returned from the applications
and returned from the services, has been extracted to a utility used
for both cases.  This should ensure that they stay in sync.

[#61931886] [Finishes #3]