Prefer -dev services
The http://run.pivotal.io instance of Cloud Foundry has '-dev'
representations of all of the services exposed through the
marketplace.  In an effort to better identify our test usage, we want
to use these whenever they exist.  This change causes service creation
to prefer the use of '-dev' services if they are available, otherwise
fall back to using the standard service.

[#61402050]