Support use of built-in buildpack
Previously, if no buildpack was specified via CF_BUILDPACK, the
buildpack from https://github.com/cloudfoundry/java-buildpack was
used.  This prevented the system test from running against the built-
in buildpack.  This change modifies that default, so that if
CF_BUILDPACK is not specified, the built-in buildpack is used instead.

[#62696164]