Display output during tests
This change updates the test build configuration to display the output
of STDOUT and STDERR, in realtime, during the tests.  Since the tests
are run in parallel, this isn't an especially readable output and will
probably be disabled once the flaw that leaves applications around
during test failure, is fixed.