Handle push failures properly
Previously, most errors dealing with Cloud Foundry were handled
properly.  However, the pushing and starting of an application was
lumped in with application creation.  This meant that if push or start
failed, the application would not be cleaned up properly.  This change
breaks those behaviors out so that the application will be properly
cleaned up.