Clean up residual applications and services
This change adds TestRules that attempt to clean up any residual
services and applications that exist when the system test suite
launches.  It does this by analyzing the collection of applications
that exist and identifies the ones that are from system tests started
by the current user.  Those applications and services are then
deleted.  This only happens once per test suite run.

Now that the bug related to failure during application creation has
been fixed, it's unlikely to see this run as part of the CI tests.
However, it will likely run for users that abort the system test suite
during the middle of a run.  By adding this functionality, users now
have more leeway as to when they can abort a local test run.