De-parallelize builds
In order to get the residual application and service cleanups to work,
the tests cannot be run in parallel.  Since the build is still running
around 20 minutes, this is not yet a problem.  This change makes the
tests run serially again, and we can re-visit this decision if they
become too slow.

As well, the version of Spring and Boot have both been upgraded.