Discriminator configurability
Previously, the discriminator used in the name of system test
artifacts was always the name of the user running the tests.  Since it
is a possible that a single runner might run tests in parallel (e.g. a
CI server) this wasn't sufficient.  The discriminator can now be
configured with the CF_DISCRIMINATOR environment variable and defaults
to the name of the user running the tests.

[#63367398]