Print crash logs on failure
Previously, if a failure happened, only the stack trace of the test
was output.  This was insufficient in diagnosing problems in the test
applications (and their supporting libraries).  This change causes the
crash logs for the application to be logged whenever there is a
failure during startup of the application.

[#63440678]