Delete route only if it exists
Previously, during residual application cleanup, the framework would
always try to delete a route for a found application.  This could lead
to situations where residual application cleanup always failed because
the application existed, but the route that matched the application
did not.  This change ensures that a route is deleted only if it
already exists.

[#63593586]