Fix logging to show actual application being tested
Currently the bespoke method for getting the name of a test is only
called once per test type so each individual test is shown as running
on the same application when it is not. This commit makes a change to
the JUnit Description code to properly display the application being
tested in the console logging.

[#95255126]