Forward Buildpack Configuration
Previously, there was a problem when testing buildpacks that required
configuration when applications were pushed.  There was no mechanism to pass
along that configuration to the applications themselves.  This change inspects
the environment of the system test itself and if it sees any `JBP_CONFIG_*`
environment variables, it will set them on the application before starting it.