Java Client Dependency Updates
This change updates the version of the Java Client in the test framework. This
change allows the code to use the new version of the push API that reads a
manifest from a file uses that to push the application.  Using this new API
allows a significant decrease in the amount of code.