Improve Shutdown Deletion
Previously, the deletion of applications depended on a JVM shutdown hook.
Since this was both time limited and run after the application context had
been destroyed, it was common for applications to not be deleted properly.
This change adds that same functionality but as an application context
shutdown hook (@PreDestroy).  This should solve the issue of proper cleanup.