Skip Services Properly
Previously, the code attempted to skip test runs for services that were
disabled.  This worked, but assumed that you still created the service and
just didn't use it.  This change updates the code such that when a service is
disabled, the tests never attempt to use it and tests are ignored as expected.

[resolves #18]