Cleanup Fail on First Push
Currently, pushing of applications the first time takes place as part of an
@PostConstruct method and cleanup is done as part of an @PreDestroy method.
However, if there's a failure in an @PostConstruct method (e.g. when pushing
an application), then the @PostConstruct method is never called.

This change puts explicit failure handling into the @PreConstruct method in
order to ensure that failures there are still cleaned up.